mint-frontend: Initial commit
This project encapsulates a Mint frontend for NFT projects to
incorporate to their website.  It assumes that an NFT vending machine
like ``cardano-nft-vending-machine`` is monitoring the payment address,
and then performing the minting transactions to send back to the chain.

Instructions for quickstart can be found in README.md.

BEFORE USING THIS PLEASE VALIDATE ALL ADDRESSES AND TRANSACTIONS.  THIS
SOFTWARE IS PROVIDED COMPLETELY AS IS WITH NO WARRANTY WHATSOEVER.

[ Documentation: README.md ]
[ Testing: Manual using ``static-server`` ]