Wallet: Enable multi-wallet support (following CIP-30)
Currently, the Lucid library offers support for three different wallet
types: Nami, Eternl (formerly CCVault) and Flint.  CIP-30 (see
https://www.cardano-caniuse.io/) has support for more, but we don't
recommend them here as we don't know how it will interact with the Lucid
library code.

This change creates a drop-down option for the "connect-wallet" menu and
enables users to select which of the three options they would like to
connect.  That selection is then cached in the Javascript module's
memory so that all future payments can run through the selected wallet.

NOTE: this does not contemplate any mobile app support.  It appears that
Eternl has support on mobile but no support for Eternl is seen in the
documentation.

NOTE: This abides by the "boy scout" rule of cleaning up as we go along,
which is why there are so many CSS changes.

[ Documentation: None ]
[ Testing: Manual enablement and transaction sending on Testnet ]