Refactoring: Build out NftToolkit library structure
We are going to turn this repository into a generic library with several
modules packed into it capable of being added to any dApp (wallet
selectors, NFT creators, etc.).  But in order to do that, we need to
create a little structure around the package and naming clarity to avoid
some overloaded gargantuan JS file.  This change introduces the
following structure:

-- src/
   -- js/
      -- library.js [top-level importer]
      -- tools/
         -- drop-payment.js [formerly "mint-frontend.js"]
         -- wallet-selector.js [picker portions of "mint-frontend.js"]
         -- module2.js [whatever is next...]
   -- static/
      -- html/
         -- webpage1.html [etc. etc.]
      -- css/
         -- stylesheet1.css [etc. etc.]

[ Documentation: README.md ]
[ Testing: Manual clean of dist/ folder then rebuild and manual QA ]