nft-creator: Provide support for gatekeeping on Mainnet using Secrets
This change uses Lucid (mainnet only) to retrieve the UTXOs that contain
a specific policy ID and validate that there are a minimum amount
required.  Note that we require a Secrets constant and not a parameter
to avoid easy swap attacks where methods are called from the console
with separate parameters.

Testnet remains unguarded.

[ Documentation: <Add wiki links, documentation not already in code> ]
[ Testing: <What automated (unit, integ) and manual testing was done?> ]