Vending Machine: Gatekeep based on number of required NFTs owned
This change provides a function for validating both at metadata file
upload time and when the vending machine is initialized, that the user
owns the requisite NFTs.  For individuals compiling this from source and
running their own local vending machine, the ``secrets.js`` file can be
updated with your gatekeeping NFT or the validation can be removed
altogether.

[ Documentation: None ]
[ Testing: Manual (mainnet and testnet) ]