nft-creator: Support multiple minting
There is native support for minting multiple assets at a time, but it's
not perfectly clean.  There appear to be errors in the library when
submitting this transaction because the transaction builder does not
correctly balance.  So, we bring back our trusty RebateCalculator class
(yay!).  With this, we can calculate exactly how much Ada to go back to
the user's wallet.  Since this is going to the same address, it matters
less that we make this the lowest amount possible.

NOTE: We remove the "self-payment" of required policy assets because not
only is it a little sketchy to send off required assets (we can use
reference inputs later), it's causing issues with balancing on testnet
(and is not even used in vending machine).

[ Documentation: None, UI changes ]
[ Testing: Manual and deploy to squarespace ]