Sweep: Support arbitrary pay-to transactions (enables bulk purchase)
The `pay_to_address` txn type was not flexible enough to enable a bulk
purchase (which required datums, redeemers, a validator, and TTL).  So,
we add those options inside of the txn schema itself here while
maintaining backwards compatibility with the existing serial purchaser.

Note: We include a jpgstore.js file specifying certain utilities for
getting datums and redeemers but it is all a bit hacky.  Also, it does
not support multiple marketplace purchases.  At the moment, this remains
in prototype state.

[ Documentation: None, handles txns seamlessly ]
[ Testing: Bulk purchase against JPG.Store ]