Lucid: Upgrade to 0.6.4 for Vasil
The Vasil HFC occurred on 9/22/22 so we are now ready to enter the
Babbage Era and fully upgrade our stack to the newest version of Lucid.
However, there are some issues with determining the testnet network
(because now there are three types of testnets, all of which return 0 as
the network ID per CIP-30).  So, we include a few fixes during our
upgrade to enable these new network keys.

NOTE: We lose some of the strict testnet enforcement as we do not need
to trap those who might inspect the minified JS code.  If they can
really figure all of this out they don't need to be gatekept since all
of this is open-source in the first place.

[ Documentation: None, UI does not contemplate testnet ]
[ Testing: Full manual QA on both pages (nft-creator and vending-machine) ]