Voting: Dummy ballot and voting power message
This change reads in the UTxOs from your wallet to determine how much
voting power you actually have.  Note that this is technically insecure
since browser-based wallets can be spoofed, but it's just a user nicety.
The smart-contract ensures the reference NFTs were sent-to-self in a
transaction, which means no one can vote without ownership of the
reference NFTs.

NOTE: The print() calls in the smart contract were commented out due to
memory constraints.

[ Documentation: None (simple UI) ]
[ Testing: Manual QA ]