Voting: Send output to a simple vote counting script
The simple vote counting script simply checks that the Oracle/vote
counter has signed the transaction.  While this gives the vote counter
free reign for now, we anticipate a future version iterating over the
outputdatums and ensuring that the UTxO assets are returned to the user.

NOTE: We add a simple output viewer of the JSON vote representation.
The address is only drained of UTxOs if the user follows through on the
transaction that appears at the end.

[ Documentation: None ]
[ Testing: Manual running E2E ]