Voting: Allow for additional metadata to be added to the ballot
This could be potentially made optional, but meh.  We want to force the
user to provide some nice CIP-0025 additions to ensure that basic voters
can receive a nice commemorative NFT (kinda like an "I, Voted!"
sticker).

NOTE: This does not impact the chain voting mechanics AT ALL and
therefore we should communicate to users that metadata does not
necessarily show how the final vote was cast.

[ Documentation: None ]
[ Testing: Manual QA (minting and viewing on pool.pm/test/metadata) ]