Voting: Helios minting policy optimization
Before we were running into mem/CPU constraints that limited us to
creating 9 ballots at a time.  This change upgrades Helios to a more
memory/cpu-efficient version to relieve the constraints.  In addition,
we use the new `map_keys` and `Value::from_map` functions to reduce the
O(n^2) looping operation with some pre-computation.

Now we can scale up to 26 assets (but set the limit at 20 in case the
UTxOs are messy or there are other transaction constraints for real
voters).

[ Documentation: None ]
[ Testing: Manually minted 20 at once ]