Bulk Listing: Initial implementation of prototype for bulk listing
This is a simple one-module prototype in JavaScript (HTML is found in
the Wild Tangz squarespace library) that allows a user to bulk list
their NFTs.

Most of the code here is used to modify the user interface and split up
UTxOs.  As such, we try to re-use functions from the Sweep module
wherever possible.

[ Documentation: None ]
[ Testing: Manual QA ]