Bulk Listing: Support for bulk delisting
JPG returns the tokens that have been listed in the `listings` field of
the JSON we are already getting so we just merge them into the existing
DOM UI and update the listing object throughout the codebase to have a
new `type` attribute.  There is some custom sniffing going on throughout
the code but since Javascript is basically duck-typed JSON, there's no
easy way around this.

[ Documentation: None ]
[ Testing: Manual bulk delisting and joint listing/delisting ]