Multi-Sig Mints: Provide Demo for Multi-Sig Mint Arch
This change makes a simple web-page with your very own "Build-A-Tangz"!
The goal is to show how changing different parts of the metadata can
impact the user's creation of the transaction, then using that to create
a witness set so that another party could complete the multi-sig mint.

Note that this does not actually do anything with the metadata and does
not have a separate server for the second half of the mint (signing with
the policy key).  There is also no validation here, which any minter
should perform to ensure they are not signing a bad action.

We recommend lift-and-shift of teh cardano-nft-js-toolkit library into
some sort of serverless NPM stack so that the user can hit a centralized
endpoint that validates the proposed mint and hides any sensitive keys.

[ Documentation: None ]
[ Testing: Multi-sig mint through the page ]